🛠 refactor sign in package(#30)
