+++ app/src/test/java/com/mashup/SignInValidationTest.kt
@@ -1,7 +1,7 @@
 package com.mashup
 
-import com.mashup.ui.register.ui.signin.fragment.validationId
-import com.mashup.ui.register.ui.signin.fragment.validationPwd
+import com.mashup.ui.signin.fragment.validationId
+import com.mashup.ui.signin.fragment.validationPwd
 import junit.framework.Assert.assertEquals
 import org.junit.Test
